Course6-API/LittleLemon -/pyproject.toml
[tool.poetry]
name = "littlelemon"
version = "0.1.0"
description = ""
authors = ["Valdielen Casarin <valcanterellic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-xml = "^2.0.0"
djangorestframework-yaml = "^2.0.0"
djangorestframework-csv = "^3.0.2"
sql = "^2022.4.0"
mysqlclient = "^2.2.4"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
